[v - a 0.6.9 (nice) ] fixed bug where frame was a long double instead of a unsigned long long

+++ ConsoleApplication2/GameState.h
@@ -4,7 +4,7 @@
 namespace Skele_lib {
 	namespace SKGE {
 		struct World {
-			long double frame;
+			unsigned long long frame;
 			long double fElapsedTime;
 			int TFPS;
